refactor: remove allocation and obligation from payments
