+++ core/credit/src/credit_facility/mod.rs
@@ -18,6 +18,7 @@ use crate::{
         CreditFacilityInterestAccrualCycle, CreditLedger,
     },
     obligation::Obligations,
+    payment_allocation::*,
     primitives::*,
     terms::InterestPeriod,
 };
@@ -40,6 +41,7 @@ where
 {
     repo: CreditFacilityRepo<E>,
     obligations: Obligations<Perms, E>,
+    payment_allocation_repo: PaymentAllocationRepo<E>,
     authz: Perms,
     ledger: CreditLedger,
     price: Price,
@@ -59,6 +61,7 @@ where
             ledger: self.ledger.clone(),
             price: self.price.clone(),
             governance: self.governance.clone(),
+            payment_allocation_repo: self.payment_allocation_repo.clone(),
         }
     }
 }
@@ -113,6 +116,8 @@ where
             .init_policy(crate::APPROVE_CREDIT_FACILITY_PROCESS)
             .await;
 
+        let payment_allocation_repo = PaymentAllocationRepo::new(pool, publisher);
+
         Self {
             repo,
             obligations: obligations.clone(),
@@ -120,6 +125,7 @@ where
             ledger: ledger.clone(),
             price: price.clone(),
             governance: governance.clone(),
+            payment_allocation_repo,
         }
     }
 
@@ -333,6 +339,81 @@ where
             facility_accrual_cycle_data: (accrual_cycle_data, credit_facility.account_ids).into(),
             new_cycle_data,
         })
+    }
+
+    #[instrument(
+        name = "credit.credit_facility.record_allocation_in_op",
+        skip(self, db),
+        err
+    )]
+
+    pub async fn record_allocation_in_op(
+        &self,
+        db: &mut es_entity::DbOp<'_>,
+        credit_facility_id: CreditFacilityId,
+        payment_id: PaymentId,
+        amount: UsdCents,
+        effective: impl Into<chrono::NaiveDate> + std::fmt::Debug + Copy,
+        audit_info: &audit::AuditInfo,
+    ) -> Result<Vec<PaymentAllocation>, CreditFacilityError> {
+        let res = self
+            .obligations
+            .allocate_payment_in_op(
+                db,
+                credit_facility_id,
+                payment_id,
+                amount,
+                effective.into(),
+                audit_info,
+            )
+            .await?;
+
+        let allocations = self
+            .payment_allocation_repo
+            .create_all_in_op(db, res.allocations)
+            .await?;
+
+        let amount_allocated = allocations.iter().fold(UsdCents::ZERO, |c, a| c + a.amount);
+        tracing::Span::current().record(
+            "amount_allocated",
+            tracing::field::display(amount_allocated),
+        );
+
+        Ok(allocations)
+    }
+
+    pub(super) async fn find_allocation_by_id_without_audit(
+        &self,
+        payment_allocation_id: impl Into<PaymentAllocationId> + std::fmt::Debug,
+    ) -> Result<PaymentAllocation, CreditFacilityError> {
+        let allocation = self
+            .payment_allocation_repo
+            .find_by_id(payment_allocation_id.into())
+            .await?;
+
+        Ok(allocation)
+    }
+
+    #[instrument(name = "core_credit.payment.find_allocation_by_id", skip(self), err)]
+    pub async fn find_allocation_by_id(
+        &self,
+        sub: &<<Perms as PermissionCheck>::Audit as AuditSvc>::Subject,
+        payment_allocation_id: impl Into<PaymentAllocationId> + std::fmt::Debug,
+    ) -> Result<PaymentAllocation, CreditFacilityError> {
+        let payment_allocation = self
+            .payment_allocation_repo
+            .find_by_id(payment_allocation_id.into())
+            .await?;
+
+        self.authz
+            .enforce_permission(
+                sub,
+                CoreCreditObject::credit_facility(payment_allocation.credit_facility_id),
+                CoreCreditAction::CREDIT_FACILITY_READ,
+            )
+            .await?;
+
+        Ok(payment_allocation)
     }
 
     pub async fn find_by_id_without_audit(
